[GOBBLIN-1209] Provide an option to configure the java tmp dir to the …
Closes #3056 from sv2000/javaTmpDir
